organized code into folders
