--- src/components/Nineties.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-
-export default function Nineties() {
-    return (
-        <div>
-            90s Television 
-            
-
-
-
-        </div>
-    )
-
-
-
-}
