+++ src/components/home/VideoPlayer.jsx
