+++ src/components/splashPage/SplashPage.jsx
